platforms/aws/requirements.txt
# AWS Lambda (boto3はLambda環境にプリインストール済み)
# boto3>=1.34.0  # Lambda環境に含まれる

# LangChain framework
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.2.0

# Cloud LLM providers
langchain-openai>=0.1.0              # Azure OpenAI (if needed)
langchain-google-vertexai>=1.0.0     # GCP Vertex AI (if needed)
langchain-aws>=0.1.0                 # AWS Bedrock (Claude)

# Utilities
python-dotenv>=1.0.0

# Document processing
pypdf>=3.0.0
openpyxl>=3.1.0
python-docx>=1.0.0

# ==============================================================================
# OCR Providers
# ==============================================================================

# AWS Textract (recommended for AWS) - boto3で使用可能
# 追加パッケージ不要

# Azure Document Intelligence (optional)
# azure-ai-documentintelligence>=1.0.0
# azure-core>=1.30.0

# Tesseract OCR (optional - Lambda Layerで対応可能)
# pytesseract>=0.3.10
# Pillow>=10.0.0
